requirements.txt
pytest
allure-pytest==2.14.0
selene==2.0.0rc9
allure-python-commons~=2.14.0